requirements.txt
numpy==1.24.2
pygame==2.1.3
Pillow~=9.4.0